MARS-1022 Move to pages instead of modals
