+++ client/src/components/WorkspaceSwitcher/index.tsx
@@ -11,15 +11,15 @@ import {
   Spacer,
   Text,
   Tooltip,
-  useDisclosure,
   useToast,
 } from "@chakra-ui/react";
 import Icon from "@components/Icon";
-import WorkspaceCreateModal from "@components/WorkspaceCreateModal";
-import WorkspaceUpdateModal from "@components/WorkspaceUpdateModal";
 
 // GraphQL resources
 import { gql, useLazyQuery, useQuery } from "@apollo/client";
+
+// Navigation
+import { useNavigate } from "react-router-dom";
 
 // Custom types
 import { IAuth, WorkspaceModel } from "@types";
@@ -31,33 +31,20 @@ import { WorkspaceContext } from "src/Context";
 
 const WorkspaceSwitcher = () => {
   const toast = useToast();
+  const navigate = useNavigate();
 
   // Store all Workspaces
   const [workspaces, setWorkspaces] = useState([] as WorkspaceModel[]);
 
-  // Value displayed on the `Select` component
-  const [labelValue, setLabelValue] = useState("");
-
   // Access token to set the active Workspace
   const [token, setToken] = useToken();
-  const [workspaceIdentifier, setWorkspaceIdentifier] = useState(
-    token.workspace,
-  );
 
   // Workspace context value
-  const { setWorkspace, setWorkspaceLoading } = useContext(WorkspaceContext);
+  const { workspace, setWorkspace, setWorkspaceLoading, label, setLabel } =
+    useContext(WorkspaceContext);
 
+  // Switcher drop-down visibility state
   const [isOpen, setIsOpen] = useState(false);
-  const {
-    isOpen: isCreateOpen,
-    onOpen: onCreateOpen,
-    onClose: onCreateClose,
-  } = useDisclosure();
-  const {
-    isOpen: isUpdateOpen,
-    onOpen: onUpdateOpen,
-    onClose: onUpdateClose,
-  } = useDisclosure();
 
   // Queries (active and lazy) to retrieve all Workspaces
   const GET_WORKSPACES = gql`
@@ -96,27 +83,21 @@ const WorkspaceSwitcher = () => {
   const updateWorkspace = async () => {
     setWorkspaceLoading(true);
 
-    // When the `workspaceIdentifier` value changes, retrieve updated model
+    // When the `workspace` value changes, retrieve updated model
     const resultWorkspace = await getWorkspace({
       variables: {
-        _id: workspaceIdentifier,
+        _id: workspace,
       },
     });
 
     if (resultWorkspace.data?.workspace) {
       setWorkspace(resultWorkspace.data.workspace._id);
-      setLabelValue(resultWorkspace.data.workspace.name);
+      setLabel(resultWorkspace.data.workspace.name);
 
       // Clone the existing token and update with selected Workspace ID
       const updatedToken: IAuth = _.cloneDeep(token);
-      updatedToken.workspace = workspaceIdentifier;
+      updatedToken.workspace = workspace;
       setToken(updatedToken);
-    }
-
-    // Refresh the list of Workspaces
-    const resultWorkspaces = await getWorkspaces();
-    if (resultWorkspaces.data?.workspaces) {
-      setWorkspaces(resultWorkspaces.data.workspaces);
     }
 
     if (workspaceError || workspacesError) {
@@ -132,6 +113,17 @@ const WorkspaceSwitcher = () => {
 
     // Update the Workspace loading state
     setWorkspaceLoading(false);
+  };
+
+  /**
+   * Utility function to update the list of Workspaces
+   */
+  const updateWorkspaces = async () => {
+    // Refresh the list of Workspaces
+    const resultWorkspaces = await getWorkspaces();
+    if (resultWorkspaces.data?.workspaces) {
+      setWorkspaces(resultWorkspaces.data.workspaces);
+    }
   };
 
   // Manage data once retrieved
@@ -161,19 +153,24 @@ const WorkspaceSwitcher = () => {
   }, []);
 
   useEffect(() => {
-    if (workspaceIdentifier !== "") {
+    if (workspace !== "") {
       updateWorkspace();
     } else {
-      onCreateOpen();
+      navigate("/create/workspace");
     }
-  }, [workspaceIdentifier]);
+  }, [workspace]);
+
+  // When the label is updated, refresh the list of Workspaces
+  useEffect(() => {
+    updateWorkspaces();
+  }, [label]);
 
   /**
    * Handle selecting a Workspace from the drop-down
    * @param _id Identifier of selected Workspace
    */
   const handleWorkspaceClick = (_id: string) => {
-    setWorkspaceIdentifier(_id);
+    setWorkspace(_id);
     setIsOpen(false);
   };
 
@@ -182,7 +179,7 @@ const WorkspaceSwitcher = () => {
    */
   const handleUpdateClick = () => {
     // Open the update Workspace modal
-    onUpdateOpen();
+    navigate(`/workspaces/${workspace}`);
 
     // Ensure `WorkspaceSwitcher` is closed
     setIsOpen(false);
@@ -193,7 +190,7 @@ const WorkspaceSwitcher = () => {
    */
   const handleCreateClick = () => {
     // Open the create Workspace modal
-    onCreateOpen();
+    navigate("/create/workspace");
 
     // Ensure `WorkspaceSwitcher` is closed
     setIsOpen(false);
@@ -208,9 +205,10 @@ const WorkspaceSwitcher = () => {
           rounded={"md"}
           border={"1px"}
           borderColor={"gray.200"}
-          bg={"white"}
-          _hover={{ bg: "gray.300" }}
+          bg={workspaces.length === 0 ? "gray.100" : "white"}
+          _hover={{ bg: workspaces.length === 0 ? "" : "gray.300" }}
           onClick={() => setIsOpen(!isOpen)}
+          disabled={workspaces.length === 0}
         >
           <Flex
             direction={"row"}
@@ -221,7 +219,7 @@ const WorkspaceSwitcher = () => {
             mr={"2"}
           >
             <Text fontSize={"sm"} fontWeight={"semibold"}>
-              {_.truncate(labelValue, { length: 15 })}
+              {_.truncate(label, { length: 15 })}
             </Text>
             <Spacer />
             <Icon name={isOpen ? "c_up" : "c_down"} />
@@ -241,19 +239,19 @@ const WorkspaceSwitcher = () => {
 
           <MenuGroup>
             {/* Create a list of all Workspaces the user has access to */}
-            {workspaces.map((workspace) => {
+            {workspaces.map((accessible) => {
               return (
                 <Tooltip
-                  key={workspace._id}
-                  isDisabled={workspace._id !== workspaceIdentifier}
+                  key={accessible._id}
+                  isDisabled={accessible._id !== workspace}
                   label={"Current Workspace"}
                 >
                   <MenuItem
-                    isDisabled={workspace._id === workspaceIdentifier}
-                    onClick={() => handleWorkspaceClick(workspace._id)}
+                    isDisabled={accessible._id === workspace}
+                    onClick={() => handleWorkspaceClick(accessible._id)}
                   >
                     <Text fontSize={"sm"} fontWeight={"semibold"}>
-                      {workspace.name}
+                      {accessible.name}
                     </Text>
                   </MenuItem>
                 </Tooltip>
@@ -291,25 +289,6 @@ const WorkspaceSwitcher = () => {
           </MenuGroup>
         </MenuList>
       </Menu>
-
-      {/* Modal to create a Workspace */}
-      <WorkspaceCreateModal
-        isOpen={isCreateOpen}
-        onOpen={onCreateOpen}
-        onClose={onCreateClose}
-        workspaceIdentifier={workspaceIdentifier}
-        setWorkspaceIdentifier={setWorkspaceIdentifier}
-        setWorkspaces={setWorkspaces}
-      />
-
-      {/* Modal to update the current Workspace */}
-      <WorkspaceUpdateModal
-        isOpen={isUpdateOpen}
-        onOpen={onUpdateOpen}
-        onClose={onUpdateClose}
-        workspaceIdentifier={workspaceIdentifier}
-        onUpdate={updateWorkspace}
-      />
     </Flex>
   );
 };
