+++ client/src/index.tsx
@@ -43,7 +43,25 @@ const authLink = setContext((_, { headers }) => {
 
 const client = new ApolloClient({
   link: ApolloLink.from([authLink, httpLink as unknown as ApolloLink]),
-  cache: new InMemoryCache({ addTypename: false }),
+  cache: new InMemoryCache({
+    typePolicies: {
+      Workspace: {
+        keyFields: ["_id"],
+      },
+      Entity: {
+        keyFields: ["_id"],
+      },
+      Project: {
+        keyFields: ["_id"],
+      },
+      Attribute: {
+        keyFields: ["_id"],
+      },
+      Activity: {
+        keyFields: ["_id"],
+      },
+    },
+  }),
 });
 
 // Render the application
