+++ types/index.d.ts
@@ -164,27 +164,6 @@ export type WorkspaceModel = IWorkspace & {
   _id: string;
 };
 
-export type WorkspaceCreateProps = {
-  // Modal state
-  isOpen: boolean;
-  onOpen: () => void;
-  onClose: () => void;
-  // Workspace state
-  workspaceIdentifier: string;
-  setWorkspaceIdentifier: React.Dispatch<React.SetStateAction<string>>;
-  setWorkspaces: React.Dispatch<React.SetStateAction<WorkspaceModel[]>>;
-};
-
-export type WorkspaceUpdateProps = {
-  // Modal state
-  isOpen: boolean;
-  onOpen: () => void;
-  onClose: () => void;
-  // Workspace state
-  workspaceIdentifier: string;
-  onUpdate: () => Promise<void>;
-};
-
 // Entity types
 export type IEntity = {
   name: string;
